requirements.txt
albumentations==1.3.1
autodistill==0.1.28
autodistill_grounding_dino==0.1.4
autodistill_yolov8==0.1.4
deep_sort_realtime==1.3.2
defusedxml==0.7.1
easyocr==1.7.1
groundingdino.egg==info
imageio==2.31.4
lightning==2.1.3
matplotlib==3.7.2
memory_profiler==0.61.0
numpngw==0.1.3
numpy==1.24.4
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
pandas==2.0.3
Pillow==10.3.0
psutil==5.9.0
psutil==5.9.8
pybboxes==0.1.6
pytorch_lightning==2.1.3
rf_segment_anything==1.0
scipy==1.13.1
segment_anything==1.0
segmentation_models_pytorch==0.3.3
supervision==0.4.0
tensorboard==2.14.0
torch==1.13.1+cu117
torchvision==0.14.1+cu117
tqdm==4.65.0
ultralytics==8.1.47
